requirements.txt
python-dotenv
pymongo
pydantic
langchain
langchain-google-genai
langgraph
chromadb
sentence-transformers